Moved required function to helpers.py will be helpfull for other user's for creating resources
